requirements.txt
numpy==1.24.3
pandas==2.0.1
python_binance==1.0.17
scikit_learn==1.2.2
python_dotenv
